# SmartRoute Extension for Hocuspocus: Revolutionary Document Routing Solution