--- a/packages/extension-smartroute/src/index.ts
+++ b/packages/extension-smartroute/src/index.ts
@@ -1,0 +1,22 @@
+export { SmartRoute } from './SmartRoute.ts';
+export { ConsistentHash } from './ConsistentHash.ts';
+export { HealthChecker } from './HealthChecker.ts';
+export { RouteManager } from './RouteManager.ts';
+
+export type { 
+  Node,
+} from './ConsistentHash.ts';
+
+export type {
+  HealthCheckOptions,
+  HealthCheckResult,
+} from './HealthChecker.ts';
+
+export type {
+  RouteManagerOptions,
+  DocumentRoute,
+} from './RouteManager.ts';
+
+export type {
+  SmartRouteConfiguration,
+} from './SmartRoute.ts';