--- a/tests/utils/redisConnectionSettings.ts
+++ b/tests/utils/redisConnectionSettings.ts
@@ -1,4 +1,4 @@
 export const redisConnectionSettings = {
   host: process.env.REDIS_HOST || '127.0.0.1',
-  port: parseInt(process.env.REDIS_PORT || '', 10) || 6379,
+  port: Number.parseInt(process.env.REDIS_PORT || '', 10) || 6379,
 }