… docs/server/extensions/smartroute.md
… package.json
… packages/extension-smartroute/README.md
… packages/extension-smartroute/package.json
… packages/extension-smartroute/src/SmartRoute.ts
… playground/backend/package.json
… playground/backend/src/smartroute-cluster.ts
… tests/extension-smartroute/ConsistentHash.ts
… tests/extension-smartroute/HealthChecker.ts
… tests/extension-smartroute/Performance.test.ts
… tests/extension-smartroute/RouteManager.ts
… tests/extension-smartroute/SmartRoute-vs-Redlock.test.ts
… tests/extension-smartroute/SmartRoute.ts